requirements.txt
fastapi
celery[redis]
redis
itsdangerous
python-multipart
jinja2
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
flower

pydantic>=2.7
pydantic-settings>=2.3
httpx>=0.27
aiofiles>=23.2
prometheus-client>=0.20
sentry-sdk>=2.6
email-validator>=2
gunicorn>=22
uvicorn[standard]>=0.30
psycopg2-binary>=2.9
sqladmin>=0.16
alembic>=1.13
